Code Crusaders/backend/requirements.txt
smartpy==0.7.0
flask==2.0.1
flask-cors==3.0.10
flask-restful==0.3.9
flask-swagger-ui==3.36.0
flask-swagger==0.2.14
flask-wtf==0.15.1
marshmallow==3.13.0
marshmallow-sqlalchemy==0.26.1
psycopg2-binary==2.9.1
sqlalchemy==1.4.25

requests==2.26.0
python-dotenv==0.19.1